optimize the code and fix sonar problem (#3413)
* optimize the code and fix sonar problem

* solve sonar bug

+++ framework/src/main/java/org/tron/core/zen/address/KeyIo.java
@@ -79,7 +79,7 @@ public class KeyIo {
       bits += fromBits;
       while (bits >= toBits) {
         bits -= toBits;
-        out.write((int) ((acc >>> bits) & maxv));
+        out.write((acc >>> bits) & maxv);
       }
     }
     if (pad) {
